LayoutTests/fast/dom/shadow/shadow-element.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<div>
  <shadow id="shadow"></shadow>
</div>
<script>
 description("This test ensures that the shadow element is available.")
var shadow = document.getElementById("shadow");
shouldBe("shadow.nodeName", "'SHADOW'");
shouldBeTrue("0 <= shadow.constructor.toString().indexOf('HTMLShadowElement')");
shouldBe("window.HTMLShadowElement", "shadow.constructor");

</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
